Make table type a dependent arrow
Even though we support dependent index sets now, all the dependencies
had to be eliminated before the final table has been created. This is
because only function arrow allowed lifting arguments to the type level.
This patch generalizes the existing table type to allow earlier indices
to appear in index spaces of latter dimensions.

Note that there is still no support for this feature in the parser and
in the codegen. Those will be implemented later.