Add a builtin Either type (and extend the pattern language)
So far there's support for the new type and patterns in the parser as
well as type inference and checking. Any attempt to actually use those
types will fail at normalization step which is not implemented yet. Note
that there are no case constructors in scope, and values can be injected
into the Either type only through `%left(expr)` or `%right(expr)`.